waste_app/requirements.txt
Flask
Pillow
numpy
joblib
werkzeug
blinker
click
itsdangerous
Jinja2
scikit-learn